refactor: extract interactive subagent tools
